插入、查询客户关系，启信宝查关系，

+++ service/src/main/java/com/scrm/service/controller/LabelController.java
@@ -31,7 +31,7 @@ public class LabelController {
     @GetMapping("/queryById")
     @ResponseBody
     public Resp queryById(
-            @RequestParam(value = "id") Integer id
+            @RequestParam(value = "id") Long id
     ) {
         Label label = labelService.queryById(id);
         if (null == label) {
@@ -76,7 +76,7 @@ public class LabelController {
     @GetMapping("/delete")
     @ResponseBody
     public Resp delete(
-            @RequestParam(value = "id") Integer id
+            @RequestParam(value = "id") Long id
     ) {
         String result = labelService.delete(id);
         if (result == null) {
